LOGIN allow Microsoft IIS/SSO authentication
Author: Martin

hack to use SSO. I integrated it into the Tools class.
Basically it will ask the server for an authenticated user and
if there is one (as per configured authentication method of
the server/application) check whether this user exists in the
Mantis database. Looks dodgy, but is not. I don't know much about PHP,
but a lot about building secure enterprise applications ;))
I coded this such, that you can integrate the code also for
non-SSO scenarios. Nothing will happen then.
To activate it, simply configure Windows Authentication in IIS
or whatever provides SSO in Apache.
Only one catch --> Mantis and global SSO logins (e.g. the windows login)
must be consistent/unique. If the Windows login 'martin' finds a different
user 'martin' in the mantis database, this will be seen as authenticated
and configured for martin.

Let me know if you need more info ;))